Changed type of size from int to long long int for weak scaling management. Added readSerieToSearch function in algo.cpp to read the data to search

+++ pattern_recognition/src/parallelDynamicAlgo.cpp
@@ -61,7 +61,7 @@ long long int ParallelDynamicAlgo::compute(Data* __restrict entireTimeSeries, Da
     return duration.count();
 }
 
-void ParallelDynamicAlgo::initData(Data *data, int size, float *values)
+void ParallelDynamicAlgo::initData(Data *data, long long int size, float *values)
 {
     data->size = size;
     data->values = new float[size];
