+++ pattern_recognition/src/parallelManualAlgoNoFirstTouch.h
@@ -7,7 +7,7 @@ class ParallelManualAlgoNoFirstTouch: public ParallelManualAlgo {
     public:
         ParallelManualAlgoNoFirstTouch();
     protected:
-        void initData(Data* data, int size, float* values) override;
+        void initData(Data* data, long long int size, float* values) override;
 };
 
 #endif
