+++ pattern_recognition/src/sequentialAlgo.cpp
@@ -12,7 +12,7 @@ SequentialAlgo::SequentialAlgo() {
 }
 
 
-void SequentialAlgo::initData(Data* data, int size, float* values) {
+void SequentialAlgo::initData(Data* data, long long int size, float* values) {
     data->size = size;
     data->values = new float[size];
     for(int i = 0; i < size; i++) {
