+++ pattern_recognition/src/sequentialAlgo.h
@@ -8,7 +8,7 @@ class SequentialAlgo : public Algo {
         SequentialAlgo();
         long long int compute(Data* __restrict entireTimeSeries, Data* __restrict timeSeriesToSearch, bool printResults);
     protected:
-        void initData(Data* data, int size, float* values);
+        void initData(Data* data, long long int size, float* values);
 };
 
 #endif
